Improvement/resolve todos (#15)
* Organized model classes into packages

* Added Transcription Screen

* Removed unused methods from Disfluency interface

* Removed Disfluency interface and implementations

* Added analysis results screen with fluency panel

* Updated graddle dependency

* Added first version of per disfluency type stats

* Added missing results to disfluency type info panel

* Added AnalysisViewModel and usage in transcription and results screens

* Added session record, integration with back-end

* Updated view model

* Added analysis results integration with back-end

* Fixed session list column

* Added session item for on-process session

* Updated home screen

* Temp fix on transcription audio player

* Updated disfluency types

* Improved audio recorder sampling rate and bit rate

* Changed AnalysedWord timestamp types to Float

* Moved audioUpload api calls from ExerciseService to new AudioUploadService

* Changed logo animation to be manually drawn instead of using an animated vector

* Changed the way the min waveform size is checked when drawing live audio waveform

* Added password requirement to contain at least 8 characters, one number and one upper case

* Added auto adjustable size text component, used in bottom navigation item labels

* Changed all list items to common format

* Added therapist exercises screen and detail screen

* Added forms to therapist and visualization of created forms

* Added missing strings resources

* Changed form completion screen design

+++ app/src/main/java/com/disfluency/api/dto/AnalysedWordDTO.kt
@@ -10,7 +10,6 @@ data class AnalysedWordDTO(
     @JsonProperty("disfluency") val disfluencyList: List<DisfluencyType>
 ) {
     fun asAnalysedWord(): AnalysedWord {
-        //TODO: cabiarlo a float
-        return AnalysedWord(text, timestamp.start.toInt(), timestamp.end.toInt(), disfluencyList)
+        return AnalysedWord(text, timestamp.start, timestamp.end, disfluencyList)
     }
 }
