+++ app/src/main/java/com/disfluency/model/analysis/AnalysedWord.kt
@@ -1,9 +1,9 @@
 package com.disfluency.model.analysis
 
-data class AnalysedWord(val word: String, val startTime: Int, val endTime: Int, val disfluency: List<DisfluencyType>? = null) {
+data class AnalysedWord(val word: String, val startTime: Float, val endTime: Float, val disfluency: List<DisfluencyType>? = null) {
 
     @OptIn(ExperimentalStdlibApi::class)
-    fun isTimeInBetween(time: Int): Boolean {
+    fun isTimeInBetween(time: Float): Boolean {
         return time in startTime..<endTime
     }
 }
